Created analysis.py file for function, added docstring
